+++ src/Monitor/Monitor/ChangeLog.md
@@ -18,6 +18,19 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* New cmdlets for SQR API (Scheduled Query Rule)  
+    - New-AzScheduledQueryRuleAlertingAction
+	- New-AzScheduledQueryRuleAznsActionGroup
+	- New-AzScheduledQueryRuleLogMetricTrigger
+	- New-AzScheduledQueryRuleSchedule
+	- New-AzScheduledQueryRuleSource
+	- New-AzScheduledQueryRuleTriggerCondition
+	- New-AzScheduledQueryRule
+	- Get-AzScheduledQueryRule
+	- Set-AzScheduledQueryRule
+	- Update-AzScheduledQueryRule
+	- Remove-AzScheduledQueryRule
+	- [More](https://docs.microsoft.com/en-us/rest/api/monitor/scheduledqueryrules) information about SQR API
 
 ## Version 1.1.0
   * New cmdlets for GenV2(non classic) metric-based alert rule